+++ doc/tutorial/sequence_file/sequence_file_solution3.cpp
@@ -50,14 +50,12 @@ IIIIIHIIJJIIIII
 
 write_file_dummy_struct go{};
 
-#if !SEQAN3_WORKAROUND_GCC_96070
 //![solution]
 #include <seqan3/std/filesystem>
 #include <seqan3/std/ranges>
 
 #include <seqan3/core/debug_stream.hpp>
 #include <seqan3/io/sequence_file/all.hpp>
-#include <seqan3/range/views/move.hpp>
 
 int main()
 {
@@ -70,28 +68,13 @@ int main()
         return std::ranges::size(rec.sequence()) >= 5;
     });
 
-    // you can use a for loop
-
-    // for (auto & rec : fin | length_filter | std::views::take(2))
-    // {
-    //     seqan3::debug_stream << "ID: " << rec.id() << '\n';
-    // }
-
-    // But you can also do this to retrieve all IDs into a vector:
-    std::vector<std::string> ids = fin
-                                 | length_filter                                    // apply length filter
-                                 | std::views::take(2)                              // take first two records
-                                 | std::views::transform(&decltype(fin)::record_type::id) // select only ID from record
-                                 // this is the same as writing:
-//                                 | std::views::transform([](auto && record)
-//                                   {
-//                                       return record.id();
-//                                   })
-                                 | seqan3::views::move                              // mark ID to be moved out of record
-                                 | seqan3::views::to<std::vector<std::string>>;     // convert to container
-    // Note that you need to know the type of id (std::string)
+    // Store all IDs into a vector:
+    std::vector<std::string> ids{};
+    for (auto & rec : fin | length_filter | std::views::take(2))
+    {
+        ids.push_back(std::move(rec.id()));
+    }
 
     seqan3::debug_stream << ids << '\n';
 }
 //![solution]
-#endif // !SEQAN3_WORKAROUND_GCC_96070